+++ examples/virtual_dpad.rs
@@ -27,9 +27,13 @@ fn spawn_player(mut commands: Commands) {
     // Stores "which actions are currently activated"
     let input_map = InputMap::default().with_dual_axis(
         Action::Move,
-        // Define a virtual D-pad using four arbitrary keys.
-        // You can also use GamepadVirtualDPad to create similar ones using gamepad buttons.
-        KeyboardVirtualDPad::new(KeyCode::KeyW, KeyCode::KeyS, KeyCode::KeyA, KeyCode::KeyD),
+        // Define a virtual D-pad using four arbitrary buttons.
+        VirtualDPad::new(
+            KeyCode::KeyW,
+            KeyCode::KeyS,
+            GamepadButtonType::DPadLeft,
+            GamepadButtonType::DPadRight,
+        ),
     );
     commands
         .spawn(InputManagerBundle::with_map(input_map))